refactor(test): extract app input test helpers and modules
