+++ src/app/input/tests/mod.rs
@@ -1,0 +1,7 @@
+mod browser_wheel;
+mod errors;
+mod helpers;
+mod keyboard;
+mod navigation;
+mod preview;
+mod preview_wheel;
